Type terrain cell asset tables

+++ include/game/track.h
@@ -551,7 +551,7 @@ extern Vec4 g_StaticSceneryPos;
  * 32x32 cell grid (clut index in the low 10 bits) and the per-cell
  * visibility rows read by GetCellVisibility. */
 extern u16 *g_TerrainCellGrid;
-extern u8 *g_CellVisibilityTable;
+extern u32 *g_CellVisibilityTable;
 
 /*
  * Their byte sizes, which is all InstallTerrainCellData needs them for - it
